pyproject.toml
[project]
name = "attension-keiba"
version = "0.1.0"
description = "Attention-based horse racing prediction MVP with backtesting and analytics."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Attension Keiba Team", email = "mlops@example.com"}]
dependencies = [
    "beautifulsoup4>=4.12",
    "numpy>=1.24",
    "pandas>=1.5",
    "requests>=2.31",
    "pyarrow>=12.0",
    "scikit-learn>=1.3",
    "torch>=2.1",
    "lightning>=2.1",
    "optuna>=3.4",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "lightgbm>=4.0",
    "tqdm>=4.66",
    "tensorboard>=2.14",
    "typer>=0.9",
    "rich>=13.7",
    "fastapi>=0.109",
    "uvicorn>=0.24",
    "joblib>=1.3",
    "plotly>=5.18"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["src"]
